+++ solver_rust_conic/totsu_core/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- Update versions of `float_eq` and `env_logger`.
 ### Deprecated
 ### Removed
 ### Fixed