+++ solver_rust_conic/totsu_f32cuda/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- Update version of `libc`.
+- Update versions of `totsu`, `float_eq` and `env_logger`.
 ### Deprecated
 ### Removed
 ### Fixed